Update to ruby/spec@0f44fd5
