+++ spec/ruby/core/proc/fixtures/refined.rb
@@ -21,31 +21,4 @@ module ProcRefinedSpecs
       end
     end
   end
-
-  # Refines operators and element access, including Hash#[] with a String
-  # key, so specs can check the specialized call paths implementations use
-  # for them.
-  module Operators
-    refine Integer do
-      def +(other)
-        "plus(#{self},#{other})"
-      end
-
-      def <(other)
-        "lt"
-      end
-    end
-
-    refine Array do
-      def [](i)
-        "at#{i}"
-      end
-    end
-
-    refine Hash do
-      def [](k)
-        "aref(#{k})"
-      end
-    end
-  end
 end
